Moved statuses and serverTypes to `SharedVars.js`
